backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pillow==11.0.0
torch==2.10.0
torchvision==0.25.0
numpy==2.4.1
opencv-python==4.10.0.84
pydantic==2.12.5
python-dotenv==1.0.0
